microservices/requirements.txt
python3 pip
flask
python-dotenv
clarifai_grpc
pandas
scikit-learn